Group By pager incorrectly interracts with SRP
We have multiple implementations of paging counters: one is regular CQL one and, the other is Group By. CQL one is counting rows as it sees them. Group By is counting rows its going to return (in other words, groups) only when it encounters the next group (or partition/data end).

The problem is that when issuing SRP, we rely on correctness of these counters. And if counter says that iterator couldn’t give enough data to exhaust limits, we’re not going to ask for more data. In case of regular counters it all works as expected, but with group by counters, we’ll say that we have seen 1 row less every time, and won’t issue SRP. In other words, if one of the nodes holds a row followed by the tombstone, current behaviour will prevent SRP from getting triggered and tombstone from being visible, since group will be unfinished, and it doesn’t make sense to ask the node for more data since it couldn’t even satisfy the current data limit.

A straightforward solution to this would have been to just start counting groups as we’re encountering them. However, here the problem is that if we say that we have a group as soon as we see the first row that belongs to it, internal pager will think that there’s no need to continue, since data limits are satisfied.

The patch is introducing a distinction between needsMoreContents and isDone . This allows us to know we need to fetch more rows in case we haven’t seen either end of data, or next row. I think we can even start counting groups right away instead of delaying counting till we see the next group to (potentially) simplify the logic, however I wanted to leave a larger refactoring to a separate patch.

Patch by Alex Petrov; reviwed by Sam Tunnicliffe and Benjamin Lerer for CASSANDRA-16307